+++ datafusion/datasource/src/mod.rs
@@ -27,6 +27,7 @@
 pub mod display;
 pub mod file;
 pub mod file_compression_type;
+pub mod file_format;
 pub mod file_groups;
 pub mod file_meta;
 pub mod file_scan_config;
